// src/generated/api.ts
export interface components {
    schemas: {
// ...
        Channel: {
            /** @enum {string} */
            channel_type: "SavedMessages";
            /** @description Unique Id */
            _id: string;
            /** @description Id of the user this channel belongs to */
            user: string;
        } | {
            /** @enum {string} */
            channel_type: "DirectMessage";
            /** @description Unique Id */
            _id: string;
            /** @description Whether this direct message channel is currently open on both sides */
            active: boolean;
            /** @description 2-tuple of user ids participating in direct message */
            recipients: string[];
            /** @description Id of the last message sent in this channel */
            last_message_id?: string | null;
        } | {
            /** @enum {string} */
            channel_type: "Group";
            /** @description Unique Id */
            _id: string;
            /** @description Display name of the channel */
            name: string;
            /** @description User id of the owner of the group */
            owner: string;
            /** @description Channel description */
            description?: string | null;
            /** @description Array of user ids participating in channel */
            recipients: string[];
            /** @description Custom icon attachment */
            icon?: components["schemas"]["File"] | null;
            /** @description Id of the last message sent in this channel */
            last_message_id?: string | null;
            /**
             * Format: int64
             * @description Permissions assigned to members of this group (does not apply to the owner of the group)
             */
            permissions?: number | null;
            /** @description Whether this group is marked as not safe for work */
            nsfw?: boolean;
        } | {
            /** @enum {string} */
            channel_type: "TextChannel";
            /** @description Unique Id */
            _id: string;
            /** @description Id of the server this channel belongs to */
            server: string;
            /** @description Display name of the channel */
            name: string;
            /** @description Channel description */
            description?: string | null;
            /** @description Custom icon attachment */
            icon?: components["schemas"]["File"] | null;
            /** @description Id of the last message sent in this channel */
            last_message_id?: string | null;
            /** @description Default permissions assigned to users in this channel */
            default_permissions?: components["schemas"]["OverrideField"] | null;
            /** @description Permissions assigned based on role to this channel */
            role_permissions?: {
                [key: string]: components["schemas"]["OverrideField"];
            };
            /** @description Whether this channel is marked as not safe for work */
            nsfw?: boolean;
            /** @description Voice Information for when this channel is also a voice channel */
            voice?: components["schemas"]["VoiceInformation"] | null;
            /**
             * Format: uint64
             * @description The channel's slowmode delay in seconds
             */
            slowmode?: number | null;
        };
        /** @description Representation of a single permission override as it appears on models and in the database */
        OverrideField: {
            /**
             * Format: int64
             * @description Allow bit flags
             */
            a: number;
            /**
             * Format: int64
             * @description Disallow bit flags
             */
            d: number;
        };
// ...
        PartialChannel: {
            name?: string | null;
            owner?: string | null;
            description?: string | null;
            icon?: components["schemas"]["File"] | null;
            nsfw?: boolean | null;
            active?: boolean | null;
            /** Format: int64 */
            permissions?: number | null;
            role_permissions?: {
                [key: string]: components["schemas"]["OverrideField"];
            } | null;
            default_permissions?: components["schemas"]["OverrideField"] | null;
            last_message_id?: string | null;
            voice?: components["schemas"]["VoiceInformation"] | null;
            /** Format: uint64 */
            slowmode?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}